% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ba_chart}
\alias{ba_chart}
\title{ba_chart}
\usage{
ba_chart(x, ...)
}
\arguments{
\item{x}{an object to be charted}

\item{...}{additional parameters}
}
\description{
A generic method to chart brapi objects
}
\examples{
if (interactive()) {
  library(brapi)
  library(magrittr)
  library(maps)

  library(brapi)

  # Need to connect to a database with genetic data

  # con <- ba_db()$xxxxx

  # ba_chart(con = con, ...)

}
}
\seealso{
Other brapiutils: \code{\link{ba_chart.ba_genomemaps_details}},
  \code{\link{ba_chart.ba_genomemaps}},
  \code{\link{ba_chart.ba_locations}},
  \code{\link{ba_describe.ba_locations}},
  \code{\link{ba_describe}}
}
\author{
Reinhard Simon
}
